Fix handling for SSH tests
We were only making an SSH connection from the host machine to the
source node, but this does not make sense. We need to test the connection
from the source to destination nodes. We accomplish this with nc.

Also add tests for the shell string to make sure it is correctly
generated in every case.